Removed redundant "View [Model]" from [Model]Actions menu if it's already possible to click on the [Model]Item to view & the [Model]Container is only used in the Show page.
Hence, only two "View [Model]" remain:
- "View Deck", since the DeckContainer may appear on the Scores page, DeckMaster, etc., & moreover the DeckFlashcard does not navigate on click.
- "View Sentence", because SentenceItem is not directly clickable when Terms are links or otherwise (see below).

Only show [Model]Actions menu for admins if the only non-admin action was "View [Model]". In the case of SentenceItem, since it is the only action for non-admins, the ContextActions menu is replaced by a button that navigates directly to the model, reducing by one the number of clicks necessary for the intended action. SentenceItem was also made always non-clickable to enforce uniformity in how it's viewed; this also allows the text to be easily selectable & scrollable.

While this change reduces redundancy & creates more intuitive navigation, it does reduce uniformity among models in terms of how they are viewed: TermItem, DeckItem & DialogItem are clickable; SentenceItem is viewable through an ad hoc button; DeckItem maintains "View Deck" in the DeckActions menu.